style: cast the gridDisk origin like every other cell wrapper
`gridDisk` was the one wrapper passing a `bigint` without the `as UInt64`
cast. Type-level only, so nothing about the emitted code changes.

+++ packages/react-native-h3/src/traversal.ts
@@ -16,7 +16,7 @@ import type { CoordIJ } from './types'
  */
 export function gridDisk(origin: bigint, k: number): BigUint64Array {
   try {
-    return new BigUint64Array(native.gridDisk(origin, k))
+    return new BigUint64Array(native.gridDisk(origin as UInt64, k))
   } catch (error) {
     rethrowAsH3Error(error)
   }
